Refactor navbar to improve code readability, add missing overflow style and fix use of useEffect on useScrollStatus
