+++ src/hooks/useScrollStatus.ts
@@ -1,0 +1,16 @@
+import { useEffect, useState } from 'react'
+
+export function useScrollStatus(threshold = 300) {
+  const [isScrolling, setIsScrolling] = useState(false)
+
+  useEffect(() => {
+    const handleScroll = () => setIsScrolling(window.scrollY > threshold)
+
+    window.addEventListener('scroll', handleScroll)
+    return () => window.removeEventListener('scroll', handleScroll)
+  }, [threshold])
+
+  return {
+    isScrolling,
+  }
+}
