+++ src/styles/globals.scss
@@ -46,3 +46,7 @@ a {
   color: inherit;
   text-decoration: none;
 }
+
+.overflow-hidden {
+  overflow: hidden;
+}
